+++ com.unity.mobile.android-logcat/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.?-?] - 2019-??-??
+
+### Fix & Improvements.
+- Added Open Terminal button
+- Use monospace font for displaying log messages, this makes text align properly when displaying addresses
+- Fix performance issues whene there's no Android device attached, the device querying will happen on worker thread.
+
 ## [0.2.7-preview] - 2019-04-24
 
 ### Minor fixes.